fix bug in TwoAddressInstruction pass regarding of reMat two addr instr

+++ java/backend/codegen/MachineInstrBuilder.java
@@ -168,7 +168,7 @@ public final class MachineInstrBuilder {
   }
 
   public MachineInstrBuilder addConstantPoolIndex(int idx,
-                                                  int offset,
+                                                  long offset,
                                                   int targetFlags) {
     mi.addOperand(MachineOperand.createConstantPoolIndex(idx, offset, targetFlags));
     return this;
